requirements.txt
is-wire==1.2.1
is-msgs==1.1.18
numpy==1.18.5
opencensus-ext-zipkin==0.2.1
python-dateutil==2.8.0
vine==1.3.0
opencv-python==4.5.4.60
face-recognition==1.3.0
scipy